bugfix: integer division in interpolation grid

+++ zrad/resize_shape.py
@@ -183,19 +183,19 @@ class ResizeShape(object):
                         for n_s in arange(0, len(M)-1): # n_s slice number
                             if M[n_s] != [] and M[n_s+1] != []: #if two consecutive slices not empty - interpolate
                                 if self.round_factor == 2:
-                                    zi = np.linspace(old_gridZ[n_s], old_gridZ[n_s+1], sliceThick/0.01+1) #create an interpolation grid between those slicse
+                                    zi = np.linspace(old_gridZ[n_s], old_gridZ[n_s+1], int(sliceThick/0.01)+1) #create an interpolation grid between those slicse
                                     #round interpolation grid accroding to specified precision
                                     for gz in arange(0, len(zi)):
                                         zi[gz] = round(zi[gz],self.round_factor)
                                     #interpolate, X list of x positions of the interpolated contour, Y list of y positions of the interpoated contour , interpolation type shape returns all the points in the structure
-                                    X, Y = InterpolateROI().interpolate(self.interpolation_algorithm, M[n_s], M[n_s+1], np.linspace(0,1,sliceThick/0.01+1), 'shape')
+                                    X, Y = InterpolateROI().interpolate(self.interpolation_algorithm, M[n_s], M[n_s+1], np.linspace(0,1,int(sliceThick/0.01)+1), 'shape')
                                 elif self.round_factor == 3 :
-                                    zi = np.linspace(old_gridZ[n_s], old_gridZ[n_s+1], sliceThick/0.001+1) #create an interpolation grid between those slicse
+                                    zi = np.linspace(old_gridZ[n_s], old_gridZ[n_s+1], int(sliceThick/0.001)+1) #create an interpolation grid between those slicse
                                     #round interpolation grid accroding to specified precision
                                     for gz in arange(0, len(zi)):
                                         zi[gz] = round(zi[gz],self.round_factor)
                                     #interpolate, X list of x positions of the interpolated contour, Y list of y positions of the interpoated contour , interpolation type shape returns all the points in the structure
-                                    X, Y = InterpolateROI().interpolate(self.interpolation_algorithm, M[n_s], M[n_s+1], np.linspace(0,1,sliceThick/0.001+1), 'shape')
+                                    X, Y = InterpolateROI().interpolate(self.interpolation_algorithm, M[n_s], M[n_s+1], np.linspace(0,1,int(sliceThick/0.001)+1), 'shape')
                                 #check which position in the interpolation grid correcpods to the new slice position
                                 for i in arange(0, len(zi)):
                                     if zi[i] in new_gridZ and zi[i] not in insertedZ: #insertedZ gathers all slice positions which are alreay filled in case that slice position is on the ovelap of two slices from orignal
